+++ activerecord/test/cases/attributes_test.rb
@@ -58,7 +58,7 @@ module ActiveRecord
       data = OverloadedType.new(non_existent_decimal: 1)
 
       assert_equal BigDecimal.new(1), data.non_existent_decimal
-      assert_raise ActiveRecord::UnknownAttributeError do
+      assert_raise ActiveModel::AttributeAssignment::UnknownAttributeError do
         UnoverloadedType.new(non_existent_decimal: 1)
       end
     end